Move PrewarmResult to domain/model
Consistent with the recent CachedActivityData refactor, moved
PrewarmResult from being defined inline in PrewarmActivityDataUseCase
to its own file in domain/model where domain types belong.
